requirements.txt
python-docx>=1.0.0
python-pptx>=0.6.21
chardet>=5.0.0
